Fix deployment: inline reducer logic to index.ts to satisfy build system
