+++ server/spacetimedb/src/index.ts
@@ -16,7 +16,6 @@
  */
 
 import { schema, table, t } from 'spacetimedb/server';
-import { sendMessage } from './reducer';
 
 export const spacetimedb = schema(
   table(
@@ -40,6 +39,23 @@ spacetimedb.clientConnected((ctx) => {
 spacetimedb.clientDisconnected((ctx) => {
   console.info(`Client disconnected: ${ctx.sender}`);
 });
+
+export const sendMessage = (ctx: any, { text }: { text: string }) => {
+  if (text.length > 1000) {
+    throw new Error('Message too long');
+  }
+
+  const senderId = ctx.sender.toHexString().substring(0, 8); // Short ID
+  const timestamp = Date.now();
+
+  console.info(`Message from ${senderId}: ${text}`);
+
+  ctx.db.global_message.insert({
+    sender: senderId,
+    text: text,
+    sent_at: timestamp
+  });
+};
 
 // Reducer to send a message
 spacetimedb.reducer('send_message', { text: t.string() }, sendMessage);
